+++ src/base/Gateway.php
@@ -624,4 +624,31 @@ abstract class Gateway extends BaseGateway
         Stripe::setApiKey($this->getApiKey());
         Stripe::setApiVersion(self::STRIPE_API_VERSION);
     }
+
+    /**
+     * Sets the default payment source in Stripe for the payment source’s customer
+     *
+     * @param string $customer The Stripe Customer ID
+     * @param string $paymentMethodId The Stripe payment method ID
+     * @return bool
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function setPaymentSourceAsDefault($customer, $paymentMethodId): bool
+    {
+        $this->configureStripeClient();
+
+        try {
+            Customer::update($customer, [
+                'invoice_settings' => [
+                    'default_payment_method' => $paymentMethodId,
+                ],
+            ]);
+
+            return true;
+        } catch (Exception $exception) {
+            Craft::error('Unable to set Stripe default payment source: ' . $exception->getMessage());
+        }
+
+        return false;
+    }
 }