+++ src/gateways/PaymentIntents.php
@@ -260,11 +260,8 @@ class PaymentIntents extends BaseGateway
         try {
             $stripeCustomer = $this->getStripeCustomer($customerId);
             $paymentMethod = PaymentMethod::retrieve($sourceData->paymentMethodId);
-            $stripeResponse = $paymentMethod->attach(['customer' => $stripeCustomer->id]);
 
-            // Set as default.
-            $stripeCustomer->invoice_settings['default_payment_method'] = $paymentMethod->id;
-            $stripeCustomer->save();
+            $stripeResponse = $paymentMethod->attach(['customer' => $stripeCustomer->id]);
 
             switch ($stripeResponse->type) {
                 case 'card':
@@ -277,6 +274,12 @@ class PaymentIntents extends BaseGateway
             }
 
             $response = $stripeResponse->toJSON();
+
+            // Make it the default in Stripe if its the only one for this gateway
+            $existingPaymentSources = Commerce::getInstance()->getPaymentSources()->getAllPaymentSourcesByCustomerId($customerId, $this->id);
+            if (!$existingPaymentSources) {
+                $this->setPaymentSourceAsDefault($stripeCustomer->id, $paymentMethod->id);
+            }
 
             return new PaymentSource([
                 'customerId' => $customerId,
@@ -321,6 +324,7 @@ class PaymentIntents extends BaseGateway
         $subscriptionParameters['expand'] = ['latest_invoice.payment_intent'];
 
         $event = new SubscriptionRequestEvent([
+            'plan' => $plan,
             'parameters' => $subscriptionParameters,
         ]);
 
@@ -343,11 +347,25 @@ class PaymentIntents extends BaseGateway
     public function deletePaymentSource($token): bool
     {
         $this->configureStripeClient();
+
+        $commercePaymentSource = Commerce::getInstance()->getPaymentSources()->getPaymentSourceByTokenAndGatewayId($token, $this->id);
+
         try {
             $paymentMethod = PaymentMethod::retrieve($token);
             $paymentMethod->detach();
         } catch (Throwable $throwable) {
-            // Assume deleted.
+            // Assume already deleted.
+        }
+
+        if ($commercePaymentSource->getIsPrimary()) {
+            $paymentSources = Commerce::getInstance()->getPaymentSources()->getAllPaymentSourcesByCustomerId($commercePaymentSource->getCustomer()->id, $this->id);
+            foreach ($paymentSources as $source) {
+                // Set it to the first that is not the one being deleted
+                if ($source->token !== $token) {
+                    Commerce::getInstance()->getCustomers()->savePrimaryPaymentSourceId($commercePaymentSource->getCustomer(), $source->id);
+                    break;
+                }
+            }
         }
 
         return true;